refactor: use categoryChipColors in TransactionRow, remove redundant toLong

+++ app/src/main/java/com/smsexpensetracker/ui/screens/dashboard/SummaryCard.kt
@@ -64,7 +64,7 @@ fun SummaryCard(
                 )
             }
             Text(
-                text = formatPaisa(animatedAmount.toLong()),
+                text = formatPaisa(animatedAmount),
                 style = MaterialTheme.typography.titleLarge,
                 fontWeight = FontWeight.Bold,
                 color = color
